+++ spec/swagger_v2/api_swagger_v2_extensions_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe 'extension' do
-  include_context 'the api entities'
+  include_context "#{MODEL_PARSER} swagger example"
 
   before :all do
     module TheApi