+++ spec/swagger_v2/api_swagger_v2_hide_documentation_path_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe 'hide documentation path' do
-  include_context 'the api entities'
+  include_context "#{MODEL_PARSER} swagger example"
 
   before :all do
     module TheApi